requirements.txt
Flask==2.3.2
gunicorn==21.2.0

google-api-python-client==2.90.0
google-auth==2.22.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.0

requests==2.31.0
